debugged

+++ Assignment 2/src/Project1/guild.cpp
@@ -120,7 +120,7 @@ namespace seneca {
 			std::cout << "No guild." << std::endl;
 		}
 		else if (m_size > 0) {
-			for (size_t i = 0; i < m_size; i++) {
+			for (int i = 0; i < m_size; i++) {
 				std::cout << "    " << i + 1 << ": " << *m_character[i] << std::endl;
 			}
 		}
