+++ Assignment 2/src/Project1/team.cpp
@@ -91,12 +91,12 @@ namespace seneca {
 		if (found) {
 			delete m_character[position];
 
-			for (int i = position; i < static_cast<size_t>(m_size - 1); i++) {
+			for (int i = position; i < m_size - 1; i++) {
 				m_character[i] = m_character[i + 1];
 			}
 
-			Character** temp = new Character * [m_size - 1];
-			for (size_t i = 0; i < m_size - 1; i++) {
+			Character** temp = new Character*[m_size - 1];
+			for (int i = 0; i < m_size - 1; i++) {
 				temp[i] = m_character[i];
 			}
